fix(prompts): include question schemas in generation
